Remove CIDER and Leiningen local configuration files.
While the idea is to show how to use CIDER in a Dockerized
environment, especially in connection to clj-refactor, it's better to
remove the local configuration files and let the end user add them.

Otherwise we hard code dependecy versions that may vary from CIDER
version to CIDER version. And pretty much the same for the Dockerized
setup and path mappings.